chore: change visibility of some classes and change the directory of ExceptionDetails
