+++ Orderly.UsersService/src/Orderly.Users.API/Middlewares/ExceptionDetails.cs
@@ -1,0 +1,8 @@
+﻿namespace Orderly.Users.API.Middlewares;
+
+internal record ExceptionDetails(
+    int Status,
+    string Type,
+    string Title,
+    string Detail,
+    IEnumerable<object>? Errors);
