+++ Orderly.UsersService/src/Orderly.Users.Application/DependencyInjection.cs
@@ -1,5 +1,4 @@
 ﻿using FluentValidation;
-using FluentValidation.AspNetCore;
 using Microsoft.Extensions.DependencyInjection;
 
 namespace Orderly.Users.Application;
